Update project to use different name
There's a weird issue with Dalamud being unable to see the plugin in the
installer, which I imagine is some banning that I'm not aware of. To
that end, the rebranding is here! Partially, just enough to get the
plugin to be visible in the installer. At some point I might rename
other stuff, but not now.